Rename Hud.ts to HudElement.ts
